requirements.txt
openai
python-dotenv
pymysql
psycopg2
beautifulsoup4
google-auth
google-auth-oauthlib
google-api-python-client
colorama
numpy
tiktoken
requests
msal 